Initial play with scrape against Google Translate
Note: This is entirely for educational purposes, it was created whilst I was playing with how to use Python. I have no idea what the legal ramifications of doing this are.  Use this at your own risk, Google literally offer a paid service for translations. It's always possible Google could block this or just change the name of the class with the result. Either would stop this script from working.